split into logical pieces
